Ensure installed hooks are writable
When you install Overcommit in a Nix environment, all files in the
gem are installed as read-only, which includes the master hook file
that gets copied when installing hooks to a Git repo. As the
installed hooks remain read-only, updating the hooks later on becomes
impossible.

We fix it by ensuring the installed master hook is writable before
copying to each hook type.

I'm not sure if it should make it writable for the group. We could
look at the umask and if user and group have the same mask, make it
writable for the group as well? As the current fix works for the
simplest of cases, I'm rolling with it.